// src/app/lib/chess/games.enum.ts
export enum GameStatus {
  PENDING = 'PENDING',
  ACTIVE = 'ACTIVE',
  FINISHED = 'FINISHED',
  ABORTED = 'ABORTED',
  INVALID = 'INVALID',
  CANCELED = 'CANCELED',
}

export enum GameResult {
  WHITE_WIN = 'WHITE_WIN',
  BLACK_WIN = 'BLACK_WIN',
  DRAW = 'DRAW',
  WHITE_WIN_BY_RESIGNATION = 'WHITE_WIN_BY_RESIGNATION',
  BLACK_WIN_BY_RESIGNATION = 'BLACK_WIN_BY_RESIGNATION',
  WHITE_WIN_ON_TIME = 'WHITE_WIN_ON_TIME',
  BLACK_WIN_ON_TIME = 'BLACK_WIN_ON_TIME',
  DRAW_INSUFFICIENT_MATERIAL = 'DRAW_INSUFFICIENT_MATERIAL',
  DRAW_STALEMATE = 'DRAW_STALEMATE',
  DRAW_THREEFOLD_REPETITION = 'DRAW_THREEFOLD_REPETITION',
  DRAW_FIFTY_MOVE_RULE = 'DRAW_FIFTY_MOVE_RULE',
  DRAW_AGREEMENT = 'DRAW_AGREEMENT',
  ABORTED = 'ABORTED',
  PENDING = 'PENDING',
}

export enum GamePiece {
  PAWN = 'PAWN',
  KNIGHT = 'KNIGHT',
  BISHOP = 'BISHOP',
  ROOK = 'ROOK',
  QUEEN = 'QUEEN',
  KING = 'KING',
}

export enum Player {
  WHITE = 'WHITE',
  BLACK = 'BLACK',
}

export enum FILE {
  A = 'a',
  B = 'b',
  C = 'c',
  D = 'd',

  E = 'e',
  F = 'f',
  G = 'g',
  H = 'h',
}

export enum MoveType {
  NOT_APPLICABLE = 'NOT_APPLICABLE',
  ILLEGAL_MOVE = 'ILLEGAL_MOVE',
  WRONG_PLAYER = 'WRONG_PLAYER',
  REGULAR_MOVE = 'REGULAR_MOVE',
  CAPTURE_MOVE = 'CAPTURE_MOVE',
  PROMOTION_MOVE = 'PROMOTION_MOVE',
  EN_PASSANT_MOVE = 'EN_PASSANT_MOVE',
  LONG_CASTLE = 'LONG_CASTLE',
  SHORT_CASTLE = 'SHORT_CASTLE',
  CHECK_MOVE = 'CHECK_MOVE',
  CHECKMATE_MOVE = 'CHECKMATE_MOVE',
  RESIGNATION_MOVE = 'RESIGNATION_MOVE',
  STALEMATE_MOVE = 'STALEMATE_MOVE',
  THREEFOLD_REPETITION_MOVE = 'THREEFOLD_REPETITION_MOVE',
  FIFTY_MOVE_RULE_MOVE = 'FIFTY_MOVE_RULE_MOVE',
  AGREEMENT_MOVE = 'AGREEMENT_MOVE',
}